Broke the code into components
