+++ src/components/ourTeamSection/MemberCardModalBodyTab.tsx
@@ -4,6 +4,9 @@ import { LocaleType } from "@/types/LocaleType";
 import { SharePopover } from "../shared/SharePopover";
 import { NoImageHeart } from "../shared/NoImageHeart";
 import { MemberDataItemType } from "@/src/mockedData/membersData";
+import { MemberTools } from "./MemberTools";
+import { MemberProjectExperience } from "./MemberProjectExperience";
+import { MemberSocialLinks } from "./MemberSocialLinks";
 
 interface MemberCardModalBodyProp {
     data: MemberDataItemType;
@@ -21,19 +24,6 @@ export const MemberCardModalBodyTab = ({ data }: MemberCardModalBodyProp) => {
 
     const getTranslation = useTranslations("MemberCardModal");
     const locale = useLocale();
-
-    const memberSocialLinks = [
-        {
-            url: socialLinks.telegram,
-            image: "/images/telegramIcon.svg",
-            alt: "Telegram",
-        },
-        {
-            url: socialLinks.linkedin,
-            image: "/images/linkedin.svg",
-            alt: "Linkedin",
-        },
-    ];
 
     return (
         <div className="hidden tab:flex relative min-w-[288px] h-auto mx-4 tab:mx-6  border border-purple-stroke">
@@ -66,33 +56,7 @@ export const MemberCardModalBodyTab = ({ data }: MemberCardModalBodyProp) => {
                         </p>
                     </div>
                 </div>
-                <ul className="flex justify-between ">
-                    {memberSocialLinks.map(({ url, image, alt }, idx) => (
-                        <li
-                            key={idx}
-                            className="flex w-full h-[34px] justify-center items-center border-t border-purple-stroke [&:nth-child(1)]:border-r"
-                        >
-                            <div className="w-6 h-6">
-                                <a
-                                    href={url}
-                                    target="_blank"
-                                    rel="noopener noreferrer"
-                                    className="cursor-pointer flex justify-center items-center w-[24px] h-[24px] pc:p-3
-            icon-hover-rounded-purple icon-hover-rounded-purple:hover icon-hover-rounded-purple:focus-visible icon-hover-rounded-purple:focus-within
-            icon-hover-rounded-purple:active relative"
-                                >
-                                    <Image
-                                        src={image}
-                                        alt={alt}
-                                        width="24"
-                                        height="24"
-                                    />
-                                </a>
-                            </div>
-                        </li>
-                    ))}
-                </ul>
-
+                <MemberSocialLinks socialLinks={socialLinks} />
                 <SharePopover
                     className="absolute bottom-0 right-0"
                     trigerShowShareText={true}
@@ -131,24 +95,9 @@ export const MemberCardModalBodyTab = ({ data }: MemberCardModalBodyProp) => {
                                 {getTranslation("projectsExperience")}
                             </span>
                         </div>
-                        <ul>
-                            {projectsExperience
-                                .slice(0, 3)
-                                .map((project, index) => (
-                                    <li
-                                        key={index}
-                                        className="px-0 truncate w-[442px] dark:pc:hover:text-red pc:hover:text-redLight dark:pc:focus:text-red pc:focus:text-redLight pc:transition pc:ease-out pc:duration-300 dark:active:text-red active:text-redLight"
-                                    >
-                                        <a
-                                            target="_blank"
-                                            href={project}
-                                            className="text-base ml-2"
-                                        >
-                                            {project}
-                                        </a>
-                                    </li>
-                                ))}
-                        </ul>
+                        <MemberProjectExperience
+                            projectsExperience={projectsExperience}
+                        />
                     </li>
                     <li className="mb-4">
                         <div className="flex items-center mb-[6px]">
@@ -157,43 +106,10 @@ export const MemberCardModalBodyTab = ({ data }: MemberCardModalBodyProp) => {
                                 {getTranslation("tools")}
                             </span>
                         </div>
-                        <ul className="overflow-hidden flex flex-wrap max-h-[80px] ml-[6px] mb-10">
-                            {tools.map((tool, idx) => (
-                                <li
-                                    key={idx}
-                                    className=" m-1 border border-grey px-2 py-[4.5px] text-base "
-                                >
-                                    {tool}
-                                </li>
-                            ))}
-                        </ul>
+                        <MemberTools tools={tools} />
                     </li>
                 </ul>
             </div>
         </div>
     );
 };
-
-{
-    /* <Image
-                src="/images/bgModalMember/SubtractTopRight.png"
-                alt="background"
-                width={320}
-                height={319}
-                className="absolute top-0 right-0 z-[-1]"
-            />
-            <Image
-                src="/images/bgModalMember/SubtractCenterRight.png"
-                alt="background"
-                width={311}
-                height={163}
-                className="absolute bottom-[28px] left-0 z-[-1]"
-            />
-            <Image
-                src="/images/bgModalMember/SubtractBottomLeft.png"
-                alt="background"
-                width={209}
-                height={192}
-                className="absolute bottom-0 left-0 z-[-1]"
-            /> */
-}
